resource-parsing/python_code/requirements.txt
cachetools==5.4.0
certifi==2024.7.4
charset-normalizer==3.3.2
frpc==0.0.6
google-api-core==2.19.1
google-auth==2.32.0
google-cloud==0.34.0
google-cloud-core==2.4.1
google-cloud-monitoring==2.22.1
google-cloud-resource-manager==1.12.4
google-cloud-storage==2.18.0
google-crc32c==1.5.0
google-resumable-media==2.7.1
googleapis-common-protos==1.63.2
grpc-google-iam-v1==0.13.1
grpcio==1.65.1
grpcio-status==1.65.1
idna==3.7
proto-plus==1.24.0
protobuf==5.27.2
pyasn1==0.6.0
pyasn1_modules==0.4.0
requests==2.32.3
rsa==4.9
urllib3==2.2.2
functions-framework==3.*
